Bulk format

+++ app/src/lib/server/providers/users/remove/index.ts
@@ -6,27 +6,24 @@ import { eq } from "drizzle-orm"
  * Remove a user from the database and all associated data
  */
 export async function remove({
-    tx=db,
-    userId,
-    returning,
+	tx = db,
+	userId,
+	returning
 }: {
-    tx?: DbTransaction | typeof db,
-    userId: string,
-    returning?: ReturningSelect
+	tx?: DbTransaction | typeof db
+	userId: string
+	returning?: ReturningSelect
 }): PromisedQueryResult<typeof returning> {
+	await users.passphrase.remove({ tx, userId })
+	await tx.delete(schema.emails).where(eq(schema.emails.userId, userId))
 
-    await users.passphrase.remove({tx, userId})
-    await tx.delete(schema.emails).where(eq(schema.emails.userId, userId))
+	const query = tx.delete(schema.users).where(eq(schema.users.id, userId))
 
-    const query = tx.delete(schema.users).where(eq(
-        schema.users.id, userId
-    ))
+	// Returning?
+	if (returning) {
+		query.returning(returning)
+	}
 
-    // Returning?
-    if (returning) {
-        query.returning(returning)
-    }
-
-    // Return result
-    return await query
+	// Return result
+	return await query
 }
